testing

+++ selfdrive/car/__init__.py
@@ -62,7 +62,7 @@ def apply_std_steer_torque_limits(apply_torque, apply_torque_last, driver_torque
     apply_torque = clip(apply_torque, apply_torque_last - LIMITS.STEER_DELTA_UP,
                         min(apply_torque_last + LIMITS.STEER_DELTA_DOWN, LIMITS.STEER_DELTA_UP))
 
-  return int(round(float(apply_torque)))
+  return round(float(apply_torque))
 
 
 def apply_toyota_steer_torque_limits(apply_torque, apply_torque_last, motor_torque, LIMITS):
@@ -82,7 +82,7 @@ def apply_toyota_steer_torque_limits(apply_torque, apply_torque_last, motor_torq
                         apply_torque_last - LIMITS.STEER_DELTA_UP,
                         min(apply_torque_last + LIMITS.STEER_DELTA_DOWN, LIMITS.STEER_DELTA_UP))
 
-  return int(round(float(apply_torque)))
+  return round(float(apply_torque))
 
 
 def crc8_pedal(data):
